j1939: Fix comparison type mismatch warnings
Fix missing (recently introduced) j1939 files that have not been fixed
in commit 46895a41c524 ("Fix comparison type mismatch warnings") from
Gary Bisson.

+++ libj1939.c
@@ -53,7 +53,7 @@ static const char *libj1939_ifnam(int ifindex)
 	fetch_names();
 
 	for (lp = saved; lp->if_index; ++lp) {
-		if (lp->if_index == ifindex)
+		if (lp->if_index == (unsigned int)ifindex)
 			return lp->if_name;
 	}
 	if (cached) {
